// src/interfaces/reqResponse.tsx
export interface ReqResListPok {
    count:    number;
    next:     string | null;
    previous: string | null;
    results:  PokLink[];
}

export interface PokLink {
    name: string;
    url:  string;
}